requirements.txt
aiogram==3.4.1
openai==1.14.2
python-dotenv==1.0.1
